requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
httpx==0.27.0
python-dotenv==1.0.1
